End of day commit: Basic structure is done. Now it needs to generate larger ascii version of each number.
