+++ src/view/view_utility.rs
@@ -1,0 +1,96 @@
+use super::*;
+
+pub(crate) fn render_progress(
+    progress: &Duration,
+    area: Rect,
+    buf: &mut Buffer,
+    title: Block,
+    duration: f64,
+) {
+    let label = Span::styled(
+        format!("{}/{}", progress.as_secs(), duration),
+        Style::new().italic().bold().fg(CUSTOM_LABEL_COLOR),
+    );
+
+    let progress = progress.as_secs_f64();
+    let ratio = ((progress / duration) * 100.0).round() / 100.0;
+    if ratio > 1.0 {
+        return;
+    }
+
+    LineGauge::default()
+        .block(title)
+        .filled_style(GAUGE_COLOR)
+        .ratio(ratio)
+        .label(label)
+        .render(area, buf);
+}
+
+pub(crate) fn title_block<'a>(color: &'a Color, author: &'a str, name: &'a str) -> Block<'a> {
+    let author = Line::from(vec![
+        Span::styled(" by ", Style::default().fg(BY_COLOR)),
+        Span::from(author),
+    ])
+    .right_aligned();
+    let name = Line::from(name).centered();
+
+    Block::new()
+        .borders(Borders::NONE)
+        .padding(Padding::vertical(1))
+        .title_bottom(author)
+        .title_bottom(name)
+        .fg(*color)
+}
+
+pub(crate) fn create_table(tracks: &Vec<Audio>) -> Table {
+    let header = Row::new(["Song", "Artist", "Duration"])
+        .style(Style::new().bold())
+        .bottom_margin(1);
+
+    let rows: Vec<Row> = tracks
+        .iter()
+        .map(|item| {
+            let style = match item.is_playing {
+                true => Style::default()
+                    .fg(Color::Blue)
+                    .add_modifier(Modifier::BOLD),
+                _ => Style::default(),
+            };
+
+            //TODO: Is cloning the only way? Investigate.
+            Row::new([
+                item.name.clone(),
+                item.author.clone(),
+                item.length.to_string(),
+            ])
+            .style(style)
+        })
+        .collect();
+
+    //let footer = Row::new(["Lemon", "Lemon Tree", "000"]);
+
+    let widths = [
+        Constraint::Percentage(50),
+        Constraint::Percentage(30),
+        Constraint::Percentage(20),
+    ];
+    let table = Table::new(rows, widths)
+        //.footer(footer.italic())
+        //.style(Color::White)
+        //.row_highlight_style(Style::new().on_black().bold())
+        //.column_highlight_style(Color::Gray)
+        //.cell_highlight_style(Style::new().reversed().yellow())
+        .header(header)
+        .column_spacing(1)
+        .row_highlight_style(Style::new().fg(Color::Green))
+        .highlight_symbol("  -  ");
+    table
+}
+
+pub(crate) fn center(area: Rect, horizontal: Constraint, vertical: Constraint) -> Rect {
+    let [area] = Layout::horizontal([horizontal])
+        .flex(Flex::Center)
+        .areas(area);
+    let [area] = Layout::vertical([vertical]).flex(Flex::Center).areas(area);
+    area
+}
